some styling and a random pirate name array to mix the results up
